% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pymol_visualize}
\alias{pymol_visualize}
\title{Generate Plot Visualization with Pymol}
\usage{
pymol_visualize(raydist, pdb, type)
}
\arguments{
\item{raydist}{Tibble obtained from the execution of read_raydist.}

\item{pdb}{Input containing only the name of the 4-digit PDB file, the file will be obtained online. If there is an extension ".pdb" or full path, the file will be obtained locally.}

\item{type}{Type of visualization: rays or dots.}
}
\description{
Visualizing the interaction between molecules is crucial in studying occluded surfaces in a protein.
The 'visualize' function will generate a plot of the protein and its interactions, using either dots or rays.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=read_raydist]{read_raydist()}}
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli

Patrick Fleming
}
