% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{read_raydist}
\alias{read_raydist}
\title{Read Raydist File}
\usage{
read_raydist(file = NULL)
}
\arguments{
\item{file}{Name of raydist file. If no parameters are provided,
it will default to searching for "raydist.lst."}
}
\description{
The FIBOS software generates a file called "raydist."
This file is crucial for generating visualization using Pymol to observe
contact points between two atoms of a protein.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=pymol_visualize]{pymol_visualize()}}
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli

Patrick Fleming
}
