% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respak.R
\name{respak}
\alias{respak}
\title{Respak Calcule}
\usage{
respak(file)
}
\arguments{
\item{file}{Prot File (.srf).}
}
\description{
The OSP value is important for verifying the quality of the values
calculated by the developed package for calculating the contact
areas between the molecules of the analyzed protein.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=occluded_surface]{occluded_surface()}}
}
\author{
Carlos Henrique da Silveira

Herson Hebert Mendes Soares

João Paulo Roquim Romanelli

Patrick Fleming
}
